+++ src/App.tsx
@@ -26,33 +26,21 @@ import {
   IconButton,
   Input,
   Select,
-  Table,
-  Tbody,
-  Td,
   Text,
-  Th,
-  Thead,
-  Tr,
   VStack,
 } from '@chakra-ui/react';
 import { useRef, useState } from 'react';
 
 import { EventForm } from './components/EventForm';
-import { NOTIFICATION_OPTIONS, WEEK_DAYS } from './constants';
+import { MonthView } from './components/MonthView';
+import { WeekView } from './components/WeekView';
+import { NOTIFICATION_OPTIONS } from './constants';
 import { useCalendarView } from './hooks/useCalendarView';
 import { useEventForm } from './hooks/useEventForm';
 import { useEventOperations } from './hooks/useEventOperations';
 import { useNotifications } from './hooks/useNotifications';
 import { useSearch } from './hooks/useSearch';
 import type { Event } from './types';
-import {
-  formatDate,
-  formatMonth,
-  formatWeek,
-  getEventsForDay,
-  getWeekDates,
-  getWeeksAtMonth,
-} from './utils/dateUtils';
 
 function App() {
   const {
@@ -136,131 +124,6 @@ function App() {
     resetForm,
   };
 
-  const renderWeekView = () => {
-    const weekDates = getWeekDates(currentDate);
-    return (
-      <VStack data-testid='week-view' align='stretch' w='full' spacing={4}>
-        <Heading size='md'>{formatWeek(currentDate)}</Heading>
-        <Table variant='simple' w='full'>
-          <Thead>
-            <Tr>
-              {WEEK_DAYS.map((day) => (
-                <Th key={day} width='14.28%'>
-                  {day}
-                </Th>
-              ))}
-            </Tr>
-          </Thead>
-          <Tbody>
-            <Tr>
-              {weekDates.map((date) => (
-                <Td key={date.toISOString()} height='100px' verticalAlign='top' width='14.28%'>
-                  <Text fontWeight='bold'>{date.getDate()}</Text>
-                  {filteredEvents
-                    .filter((event) => new Date(event.date).toDateString() === date.toDateString())
-                    .map((event) => {
-                      const isNotified = notifiedEvents.includes(event.id);
-                      return (
-                        <Box
-                          key={event.id}
-                          p={1}
-                          my={1}
-                          bg={isNotified ? 'red.100' : 'gray.100'}
-                          borderRadius='md'
-                          fontWeight={isNotified ? 'bold' : 'normal'}
-                          color={isNotified ? 'red.500' : 'inherit'}
-                        >
-                          <HStack spacing={1}>
-                            {isNotified && <BellIcon />}
-                            <Text fontSize='sm' noOfLines={1}>
-                              {event.title}
-                            </Text>
-                          </HStack>
-                        </Box>
-                      );
-                    })}
-                </Td>
-              ))}
-            </Tr>
-          </Tbody>
-        </Table>
-      </VStack>
-    );
-  };
-
-  const renderMonthView = () => {
-    const weeks = getWeeksAtMonth(currentDate);
-
-    return (
-      <VStack data-testid='month-view' align='stretch' w='full' spacing={4}>
-        <Heading size='md'>{formatMonth(currentDate)}</Heading>
-        <Table variant='simple' w='full'>
-          <Thead>
-            <Tr>
-              {WEEK_DAYS.map((day) => (
-                <Th key={day} width='14.28%'>
-                  {day}
-                </Th>
-              ))}
-            </Tr>
-          </Thead>
-          <Tbody>
-            {weeks.map((week, weekIndex) => (
-              <Tr key={weekIndex}>
-                {week.map((day, dayIndex) => {
-                  const dateString = day ? formatDate(currentDate, day) : '';
-                  const holiday = holidays[dateString];
-
-                  return (
-                    <Td
-                      key={dayIndex}
-                      height='100px'
-                      verticalAlign='top'
-                      width='14.28%'
-                      position='relative'
-                    >
-                      {day && (
-                        <>
-                          <Text fontWeight='bold'>{day}</Text>
-                          {holiday && (
-                            <Text color='red.500' fontSize='sm'>
-                              {holiday}
-                            </Text>
-                          )}
-                          {getEventsForDay(filteredEvents, day).map((event) => {
-                            const isNotified = notifiedEvents.includes(event.id);
-                            return (
-                              <Box
-                                key={event.id}
-                                p={1}
-                                my={1}
-                                bg={isNotified ? 'red.100' : 'gray.100'}
-                                borderRadius='md'
-                                fontWeight={isNotified ? 'bold' : 'normal'}
-                                color={isNotified ? 'red.500' : 'inherit'}
-                              >
-                                <HStack spacing={1}>
-                                  {isNotified && <BellIcon />}
-                                  <Text fontSize='sm' noOfLines={1}>
-                                    {event.title}
-                                  </Text>
-                                </HStack>
-                              </Box>
-                            );
-                          })}
-                        </>
-                      )}
-                    </Td>
-                  );
-                })}
-              </Tr>
-            ))}
-          </Tbody>
-        </Table>
-      </VStack>
-    );
-  };
-
   return (
     <Box w='full' h='100vh' m='auto' p={5}>
       <Flex gap={6} h='full'>
@@ -297,8 +160,17 @@ function App() {
             />
           </HStack>
 
-          {view === 'week' && renderWeekView()}
-          {view === 'month' && renderMonthView()}
+          {view === 'week' && (
+            <WeekView currentDate={currentDate} events={events} notifiedEvents={notifiedEvents} />
+          )}
+          {view === 'month' && (
+            <MonthView
+              currentDate={currentDate}
+              events={events}
+              notifiedEvents={notifiedEvents}
+              holidays={holidays}
+            />
+          )}
         </VStack>
 
         {/* 일정 검색 영역 */}
